Introduce NamedProvider
