+++ src/main/java/com/amannmalik/mcp/spi/NamedProvider.java
@@ -1,0 +1,21 @@
+package com.amannmalik.mcp.spi;
+
+import java.util.Optional;
+
+public sealed interface NamedProvider<T extends DisplayNameProvider> extends Provider<T>
+        permits PromptProvider, ResourceProvider, ToolProvider {
+    default Optional<T> find(String name) {
+        if (name == null) throw new IllegalArgumentException("name required");
+        Cursor cursor = Cursor.Start.INSTANCE;
+        do {
+            Pagination.Page<T> page = list(cursor);
+            for (T item : page.items()) {
+                if (item.name().equals(name)) {
+                    return Optional.of(item);
+                }
+            }
+            cursor = page.nextCursor();
+        } while (!(cursor instanceof Cursor.End));
+        return Optional.empty();
+    }
+}
